requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.2
blinker==1.8.2
build==1.0.3
CacheControl==0.14.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
click==8.1.7
contourpy==1.2.1
crashtest==0.4.1
cryptography==41.0.4
cycler==0.12.1
Cython==0.29.37
dill==0.3.8
distlib==0.3.7
Django==5.0.6
django-admin==2.0.2
django-bootstrap5==24.2
django-excel-response2==3.0.6
django-six==1.0.5
dnspython==2.6.1
docutils==0.21.2
dulwich==0.21.6
email_validator==2.1.1
et-xmlfile==1.1.0
excel-base==1.0.4
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.18.1
filelock==3.12.2
Flask==3.0.3
flask-ngrok==0.0.25
fonttools==4.53.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
imageio==2.34.1
importlib-metadata==6.8.0
installer==0.7.0
isort==5.13.2
isoweek==1.3.3
itsdangerous==2.2.0
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.4
keyring==24.2.0
kiwisolver==1.4.5
lazy_loader==0.4
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mdurl==0.1.2
meson==1.3.2
more-itertools==10.1.0
msgpack==1.0.7
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.17
numpy==2.0.0
openpyxl==3.1.4
orjson==3.10.5
packaging==23.2
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
pipenv==2024.0.0
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pylint==3.2.3
pymongo==4.7.3
pyngrok==7.1.6
pyparsing==3.1.2
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.5.2
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
scikit-image==0.24.0
scipy==1.13.1
screen==1.0.1
SecretStorage==3.3.3
shellingham==1.5.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.0
starlette==0.37.2
tifffile==2024.6.18
TimeConvert==3.0.13
tomlkit==0.12.1
trove-classifiers==2023.9.19
twine==5.1.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.0.6
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
xlwt==1.3.0
zipp==3.17.0

djangorestframework~=3.15.2